TST: Move some tests from test_regression somewhere more specific
